Moved GetConnectionString method to DbContext class and adjusted the classes so the namespace thing doesn't push the class away from the left side of the code view
